Replace usage of Double with BigDecimal
Replaced Double with BigDecimal for better precision in financial data handling. Changes were made across the application including entities, DTOs, and services. Furthermore, created RuleTest for BigDecimal calculations in the Rule entity. This commit prevents potential precision loss issues in mathematical operations.

+++ src/main/java/at/devp/myfinance/dto/RuleOverviewDto.java
@@ -2,13 +2,15 @@ package at.devp.myfinance.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RuleOverviewDto {
   private Long id;
   private String description;
   private String from;
   private String to;
-  private Double amount;
-  private Double oldAmount;
+  private BigDecimal amount;
+  private BigDecimal oldAmount;
   private boolean isChange;
 }
