+++ src/main/java/at/devp/myfinance/dto/TransferOverviewDto.java
@@ -2,13 +2,15 @@ package at.devp.myfinance.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class TransferOverviewDto {
   private Long id;
   private String description;
   private String from;
   private String to;
-  private Double amount;
-  private Double oldAmount;
+  private BigDecimal amount;
+  private BigDecimal oldAmount;
   private boolean isChange;
 }
